+++ packages/elasticsearch/CHANGELOG.md
@@ -1,4 +1,8 @@
-﻿### 0.6.9
+### 0.7.0
+* Introducing new example type number range histogram.
+* Number now supports find best functionality.
+* Geo type now guards against 0 results in geocoder.
+### 0.6.9
 * Facet can't allow size 0 or empty, so we're sending 10 by default (as before).
 ### 0.6.8
 * Added the context property useRaw to step out of date formattings on the date example type.